Refactor - move rooms directory under admin/events and make header its own module so any module can import its functionality.
