+++ front_end/src/app/admin/admin.component.ts
@@ -1,12 +1,11 @@
 import { Component, ViewEncapsulation } from '@angular/core';
-import { HeaderService } from './header/header.service';
 import { AdminService } from './admin.service';
 
 @Component({
   selector: 'admin',
   templateUrl: 'admin.component.html',
   styleUrls: ['admin.component.scss'],
-  providers: [ AdminService, HeaderService ],
+  providers: [ AdminService ],
   encapsulation: ViewEncapsulation.None
 })
 export class AdminComponent {}
