+++ front_end/src/app/admin/events/rooms/alternate-rooms/alternate-rooms.component.scss
