+++ front_end/src/app/admin/events/rooms/alternate-rooms/alternate-rooms.component.spec.ts
@@ -1,6 +1,6 @@
 import { AlternateRoomsComponent } from './alternate-rooms.component';
 import { ActivatedRoute, ActivatedRouteSnapshot } from '@angular/router';
-import { Room } from '../../../shared/models';
+import { Room } from '../../../../shared/models';
 import { Observable } from 'rxjs';
 
 describe('AlternateRoomsComponent', () => {
