+++ front_end/src/app/admin/events/rooms/room-group-list.component.html
