+++ front_end/src/app/admin/events/rooms/room-group-list.component.scss
@@ -1,4 +1,4 @@
-@import '../../../assets/scss/_variables.scss';
+@import '../../../../assets/scss/_variables.scss';
 
 table.groups.months {
   tr {
