+++ front_end/src/app/admin/events/rooms/room-group-list.component.ts
@@ -1,10 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { Location } from '@angular/common';
-import { HeaderService } from '../header/header.service';
-import { Group, Room, Event } from '../../shared/models';
-import { AdminService } from '../admin.service';
-import { ApiService } from '../../shared/services';
+import { HeaderService } from '../../header/header.service';
+import { Group, Room, Event } from '../../../shared/models';
+import { AdminService } from '../../admin.service';
+import { ApiService } from '../../../shared/services';
 
 @Component({
   templateUrl: 'room-group-list.component.html',
