+++ front_end/src/app/admin/events/rooms/room-group.component.scss
@@ -1,5 +1,5 @@
-@import '../../../assets/scss/_bootstrap.scss';
-@import '../../../assets/scss/_variables.scss';
+@import '../../../../assets/scss/_bootstrap.scss';
+@import '../../../../assets/scss/_variables.scss';
 
 td:first-of-type {
   width: 200px;
