+++ front_end/src/app/admin/events/rooms/room.component.spec.ts
@@ -1,6 +1,6 @@
 import { Observable } from 'rxjs/Observable';
 import { RoomComponent } from './room.component';
-import { Room } from '../../shared/models';
+import { Room } from '../../../shared/models';
 
 let fixture: RoomComponent;
 let adminServiceStub: any = {
