+++ front_end/src/app/admin/events/rooms/room.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { FormControl, FormGroup } from '@angular/forms';
-import { AdminService } from '../admin.service';
-import { Room } from '../../shared/models';
+import { AdminService } from '../../admin.service';
+import { Room } from '../../../shared/models';
 
 @Component({
   selector: '.room',
