+++ front_end/src/app/admin/header/header.component.html
@@ -22,7 +22,7 @@
       <a class="nav-link">Registration</a>
     </li>
     <li class="nav-item">
-      <a class="nav-link" [routerLink]="['/admin/guest-checkin']">Guest Check In</a>
+      <a class="nav-link" [routerLink]="['/admin/new-family-registration']">New Family Registration</a>
     </li>
   </ul>
   <form class="form-inline float-xs-right">
@@ -57,3 +57,4 @@
     <button type="button" class="btn btn-secondary" disabled>Next Event</button>
   </div>
 </nav>
+
