+++ front_end/src/app/admin/header/header.module.ts
@@ -1,0 +1,26 @@
+import { NgModule } from '@angular/core';
+import { RouterModule } from '@angular/router';
+import { MomentModule } from 'angular2-moment';
+import { SharedModule } from '../../shared/shared.module';
+
+import { HeaderComponent } from './header.component';
+import { HeaderService } from './header.service';
+
+@NgModule({
+  declarations: [
+    HeaderComponent,
+  ],
+  imports: [
+    SharedModule,
+    RouterModule,
+    MomentModule,
+  ],
+  exports: [
+    HeaderComponent,
+  ],
+  providers: [
+    HeaderService,
+  ],
+})
+
+export class HeaderModule { }
