+++ front_end/src/app/admin/header/index.ts
@@ -1,0 +1,2 @@
+export { HeaderModule } from './header.module';
+
