+++ front_end/src/app/admin/new-family-registration/new-family-registration.component.html
@@ -1,6 +1,6 @@
-<header-event></header-event>
+<!-- <header-event></header-event> -->
 
-<div class="wrapper guest-checkin">
+<div class="wrapper new-family-registration">
   <div class="container-fluid">
     <div class="row">
       <div class="col-sm-12 mb-2">
